BackEnd/requirements.txt
flask
pymongo
dnspython
python-dotenv
jsonify
bson


